# docs/tonic-v4-openapi.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "https://api.publisher.tonic.com/v4"
    }
  ],
  "info": {
    "title": "TONIC. for Publishers API",
    "description": "This is the documentation for the \"TONIC. for Publishers API\". Currently this API is in beta and is subject to change. If you have any questions or feedback, please contact your Account Manager.",
    "version": "4.16.1",
    "contact": {
      "email": "support@tonic.com"
    }
  },
...
  "paths": {
...
    "/articles": {
      "get": {
        "operationId": "ArticlesGet",
        "summary": "Get all available articles",
        "description": "Returns a list of articles. The results can be filtered and sorted by a specified field and direction.",
        "deprecated": false,
        "tags": [
          "Articles"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "articleIds",
            "description": "Filter by article IDs.<br />Pass multiple values separated by commas.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": "1,2,3"
          },
          {
            "in": "query",
            "name": "title",
            "description": "Filter by campaign name (case-insensitive, min. 3 characters).",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 100
            },
            "example": "Article XYZ"
          },
          {
            "in": "query",
            "name": "domain",
            "description": "Filter by domain",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 100
            },
            "example": "mydomain.com"
          },
          {
            "in": "query",
            "name": "language",
            "description": "Filter by 2 digits language code",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 2
            },
            "example": "en"
          },
          {
            "in": "query",
            "name": "verticalIds",
            "description": "Filter by vertical IDs<br />Pass multiple values separated by commas.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": "2,3,4"
          },
          {
            "in": "query",
            "name": "verticalName",
            "description": "Filter by vertical name (case-insensitive, min. 3 characters).",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 100
            },
            "example": "Some Vertical"
          },
          {
            "in": "query",
            "name": "offerIds",
            "description": "Filter by Offer IDs<br />Pass multiple values separated by commas.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": "3,4,5"
          },
          {
            "in": "query",
            "name": "offerName",
            "description": "Filter by Offer name (case-insensitive, min. 3 characters).",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 100
            },
            "example": "Offer XYZ"
          },
          {
            "in": "query",
            "name": "countryCode",
            "description": "Filter by country code<br />Pass multiple values separated by commas.",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 2
            },
            "example": "de,ch,us,ua"
          },
          {
            "in": "query",
            "name": "orderField",
            "description": "Sort results by field.<br /><i>Available Values</i> : id, title, created, offer_name, vertical_name, country_name, country_code, language, connected_campaigns",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "id"
          },
          {
            "in": "query",
            "name": "orderOrientation",
            "description": "Sort direction: asc (A-Z) or desc (Z-A).",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "example": "asc"
            },
            "example": "desc"
          },
          {
            "in": "query",
            "name": "limit",
            "description": "The limit of returned paginated data",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 25
          },
          {
            "in": "query",
            "name": "offset",
            "description": "The offset of returned paginated data",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 15
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                },
                "example": {
                  "data": [
                    {
                      "id": 123,
                      "offer": {
                        "id": 42,
                        "name": "Offer XYZ",
                        "vertical": {
                          "id": 76,
                          "name": "Vertical XYZ"
                        }
                      },
                      "country": {
                        "code": "US",
                        "name": "United States"
                      },
                      "language": "en",
                      "title": "My awesome article",
                      "url": "https://15u102.xdzxrxbcwtv.com/articles/lorem-ipsum",
                      "created": "2024-08-28T12:29:44-07:00",
                      "connectedCampaignsCount": 0
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        }
      }
    },
...
    "/articles/{id}": {
      "get": {
        "operationId": "ArticlesIdGet",
        "summary": "Get an article",
        "description": "Get an article data",
        "deprecated": false,
        "tags": [
          "Articles"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "The article id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 999
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                },
                "example": {
                  "data": {
                    "id": 123,
                    "offer": {
                      "id": 42,
                      "name": "Offer XYZ",
                      "vertical": {
                        "id": 76,
                        "name": "Vertical XYZ"
                      }
                    },
                    "country": {
                      "code": "US",
                      "name": "United States"
                    },
                    "language": "en",
                    "title": "My awesome article",
                    "url": "https://15u102.xdzxrxbcwtv.com/articles/lorem-ipsum",
                    "created": "2024-08-28T12:29:44-07:00",
                    "connectedCampaignsCount": 0
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "delete": {
        "operationId": "ArticlesIdDelete",
        "summary": "Delete an article",
        "description": "Removes an article from the account. It disappears from the article list and is no longer offered for new campaigns. Articles that still have a campaign attached can not be deleted, a stopped campaign included: stop the campaign and it is deleted automatically about a day later, then the article can go.",
        "deprecated": false,
        "tags": [
          "Articles"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "The article id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 999
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ApiModel": {
        "title": "ApiModel",
        "type": "object"
      },
...
      "ErrorResponse": {
        "title": "ErrorResponse",
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "minimum": 400,
            "maximum": 599,
            "example": 500
          },
          "details": {
            "title": "ErrorDetails",
            "oneOf": [
              {
                "title": "ErrorDetailsSimple",
                "type": "string",
                "example": "Internal Server Error"
              },
              {
                "title": "ErrorDetailsParameter",
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "parameterName": {
                      "type": "string",
                      "example": "name"
                    },
                    "parameterLocation": {
                      "type": "string",
                      "example": "body"
                    },
                    "collectionIndex": {
                      "type": "number",
                      "description": "Index of dataset in the collection where the error occurred",
                      "example": 0,
                      "nullable": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Missing required post body parameter 'name'"
                    }
                  }
                }
              }
            ]
          },
          "requestId": {
            "type": "string"
          }
        }
      },
...
      "Response": {
        "title": "Response",
        "type": "object",
        "properties": {
          "data": {
            "title": "ResponseData",
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ApiModel"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ApiModel"
                }
              },
              {
                "type": "object"
              },
              {
                "type": "array",
                "items": {}
              }
            ]
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "pagination": {
            "type": "object",
            "nullable": true
          },
          "sorting": {
            "type": "object",
            "nullable": true,
            "properties": {
              "orderField": {
                "type": "string",
                "example": "id"
              },
              "orderOrientation": {
                "type": "string",
                "enum": [
                  "asc",
                  "desc"
                ],
                "example": "asc"
              }
            }
          }
        }
      },
...
    },
    "responses": {
...
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": 400,
              "details": "Bad Request",
              "requestId": "root=1-6335cbde-6baf0bed7a2eaecc6b1214bf"
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": 401,
              "details": "Unauthorized",
              "requestId": "root=1-6335cbde-6baf0bed7a2eaecc6b1214bf"
            }
          }
        }
      },
      "403": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": 403,
              "details": "Forbidden",
              "requestId": "root=1-6335cbde-6baf0bed7a2eaecc6b1214bf"
            }
          }
        }
      },
      "404": {
        "description": "Not Found Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": 404,
              "details": "Not Found Error",
              "requestId": "root=1-6335cbde-6baf0bed7a2eaecc6b1214bf"
            }
          }
        }
      },
      "409": {
        "description": "Conflict Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": 409,
              "details": "Conflict Error",
              "requestId": "root=1-6335cbde-6baf0bed7a2eaecc6b1214bf"
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": 500,
              "details": "Internal Server Error",
              "requestId": "root=1-6335cbde-6baf0bed7a2eaecc6b1214bf"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}